requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
requests>=2.32.0
openai>=1.50.0
pydantic>=2.9.0
python-dotenv>=1.0.0
web3>=7.0.0
eth-account>=0.13.0
x402[fastapi,evm,extensions,httpx]>=0.3.0
cdp-sdk
jinja2>=3.1.0
